requirements.txt
fastapi
uvicorn
python-multipart
twilio
openai
aiohttp
python-dotenv
requests
jinja2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
sqlalchemy
psycopg2-binary
alembic